+++ backend/src/file/file.module.ts
@@ -3,6 +3,7 @@ import { JwtModule } from "@nestjs/jwt";
 import { DownloadLogModule } from "../download-log/download-log.module";
 import { EmailModule } from "../email/email.module";
 import { ShareDomainModule } from "../share/domain/share-domain.module";
+import { StorageModule } from "../storage/storage.module";
 import { FileController } from "./file.controller";
 import { FileService } from "./file.service";
 import { LocalFileService } from "./local.service";
@@ -14,6 +15,7 @@ import { DownloadLimitGuard } from "./guard/downloadLimit.guard";
     EmailModule,
     ShareDomainModule,
     DownloadLogModule,
+    StorageModule,
   ],
   controllers: [FileController],
   providers: [FileService, LocalFileService, DownloadLimitGuard],