+++ backend/src/jobs/jobs.module.ts
@@ -2,9 +2,10 @@ import { Module } from "@nestjs/common";
 import { FileModule } from "../file/file.module";
 import { JobsService } from "./jobs.service";
 import { ConfigModule } from "../config/config.module";
+import { StorageModule } from "../storage/storage.module";
 
 @Module({
-  imports: [FileModule, ConfigModule],
+  imports: [FileModule, ConfigModule, StorageModule],
   providers: [JobsService],
 })
 export class JobsModule {}